Custom Gamepad driver with support for internal plugins Bump xbox-xcloud-player version
