+++ src/plugins/frontend/opentrack.ts
@@ -1,6 +1,5 @@
 import Application from '../../frontend/application'
-// import GammepadDriver from 'xbox-xcloud-player/src/Driver/Gamepad'
-import xCloudPlayer from 'xbox-xcloud-player'
+import { GamepadDriver } from '../../frontend/plugins'
 
 interface OpentrackPosition {
     [key: string]: number;
@@ -23,6 +22,7 @@ export class OpentrackPluginFrontend {
         pitch: 0,
         roll: 0,
     }
+    _positionTimestamp:number
 
     constructor(application:Application = undefined) {
         this._application = application
@@ -52,12 +52,14 @@ export class OpentrackPluginFrontend {
         this._application._ipc.on('opentrack-sync', (event:any, variables:any) => { 
             this._isRunning = variables.isRunning
             this._position = variables.position
+            this._positionTimestamp = variables.timestamp
 
             this.renderSettings()
         })
 
         this._application._ipc.on('opentrack-position', (event:any, variables:any) => {
             this._position = variables.position
+            this._positionTimestamp = variables.timestamp
         })
     }
 
@@ -70,14 +72,16 @@ export class OpentrackPluginFrontend {
     }
 
     onStreamStart() {
+        // on stream start
+    }
+
+    onGamepadRequest(gamepadDriver:GamepadDriver) {
         // Check if Opentrack is enabled
-        if(this._isRunning === true){
-            console.log('Opentrack inject plugin!')
-            this._application._StreamingView._streamClient._xCloudPlayerConfig = {
-                ...this._application._StreamingView._streamClient._xCloudPlayerConfig,
-                'input_driver': new OpentrackDriver(this)
-            }
+        if(this._isRunning === true && (Date.now() - this._positionTimestamp) < 5000 ){ // 5000 ms timeout
+            return this.convertToControllerInput(this._position)
         }
+
+        return {}
     }
 
     renderSettings() {
@@ -91,126 +95,8 @@ export class OpentrackPluginFrontend {
             document.getElementById('plugin_opentrack_positions').innerHTML += '<p>'+position+': <span id="plugin_opentrack_positions_'+position+'" class="settingValue">'+this._position[position]+'</span></p>'
         }
     }
-}
-
-class GamepadDriver {
-
-    _application:xCloudPlayer
-
-    _gamepads:Array<any> = []
-
-    constructor() {
-        // this._application = application
-    }
-
-    setApplication(application:xCloudPlayer) {
-        this._application = application
-    }
-
-    start() {
-        window.addEventListener("gamepadconnected", (e) => {
-
-            const gamepad = {
-                index: e.gamepad.index,
-                name: e.gamepad.id,
-                buttons: e.gamepad.buttons,
-                axes: e.gamepad.axes,
-            }
-            this._gamepads.push(gamepad)
-
-            this._application.getEventBus().emit('gamepad_connect', gamepad)
-            console.log('xCloudPlayer Driver/Gamepad.ts - Controller connected:', this._gamepads)
-        })
-
-        window.addEventListener("gamepaddisconnected", (e) => {
-            for(const gamepad in this._gamepads){
-                if(this._gamepads[gamepad].index === e.gamepad.index){
-                    const removedGamepad = this._gamepads[gamepad]
-                    this._gamepads.splice(e.gamepad.index, 1)
-
-                    this._application.getEventBus().emit('gamepad_disconnect', removedGamepad)
-                    console.log('xCloudPlayer Driver/Gamepad.ts - Controller disconnected:', this._gamepads)
-                }
-            }
-        })
-    }
-
-    stop() {
-        // console.log('xCloudPlayer Driver/Gamepad.ts - Stop collecting events:', this._gamepads)
-    }
-
-    // Use requestState() from OpentrackDriver()
-    //
-    // requestState() {
-    //     for(const gamepad in this._gamepads){
-    //         const gamepadState = navigator.getGamepads()[this._gamepads[gamepad].index]
-    //         const state = this.mapStateLabels(gamepadState?.buttons, gamepadState?.axes, this._gamepads[gamepad].index)
-
-    //         this._application.getChannelProcessor('input').queueGamepadState(state)
-    //     }
-    // }
-
-    mapStateLabels(buttons:any, axes:any, gamepadIndex=0) {
-        return {
-            GamepadIndex: gamepadIndex,
-            A: buttons[0].value,
-            B: buttons[1].value,
-            X: buttons[2].value,
-            Y: buttons[3].value,
-            LeftShoulder: buttons[4].value,
-            RightShoulder: buttons[5].value,
-            LeftTrigger: buttons[6].value,
-            RightTrigger: buttons[7].value,
-            View: buttons[8].value,
-            Menu: buttons[9].value,
-            LeftThumb: buttons[10].value,
-            RightThumb: buttons[11].value,
-            DPadUp: buttons[12].value,
-            DPadDown: buttons[13].value,
-            DPadLeft: buttons[14].value,
-            DPadRight: buttons[15].value,
-            Nexus: buttons[16].value,
-            LeftThumbXAxis: axes[0],
-            LeftThumbYAxis: axes[1],
-            RightThumbXAxis: axes[2],
-            RightThumbYAxis: axes[3]
-        }
-    }
-}
-
-class OpentrackDriver extends GamepadDriver {
-
-    _opentrack:OpentrackPluginFrontend
-
-    constructor(openTrack:OpentrackPluginFrontend) {
-        super()
-
-        this._opentrack = openTrack
-        console.log('Opentrack: Setting up gamepad driver')
-    }
-
-    requestState() {
-        // override gamepad controls when opentrack is on?
-
-        for(const gamepad in this._gamepads){
-
-            // Lets assume we are on gamepad index 0
-            let opentrackState = {}
-            if(this._gamepads[gamepad].index === 0){
-                opentrackState = this.convertToControllerInput(this._opentrack._position)
-            }
-
-            const gamepadState = navigator.getGamepads()[this._gamepads[gamepad].index]
-            const state = this.mapStateLabels(gamepadState?.buttons, gamepadState?.axes, this._gamepads[gamepad].index)
-
-            const mergedState = { ...state, ...opentrackState }
-
-            this._application.getChannelProcessor('input').queueGamepadState(mergedState)
-        }
-    }
 
     convertToControllerInput(position:any) {
-        // Run js func...
         const state = {
             RightThumbXAxis: (position.yaw/20),
             RightThumbYAxis: -(position.pitch/20)
